requirements.txt
cudf-cu12==24.02.0
cuml-cu12==24.02.0
xgboost==2.0.3

Flask==3.1.0 
gunicorn==23.0.0

pandas==2.1.4
numpy==1.26.4
scikit-learn==1.3.2
